+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use Bevy Rapier's built-in Damping component for velocity dampening
+
 ## [0.0.24] - 2023-11-15
 
 ### Added